Fix package removal for puppet 3.5.x
In puppet 3.5.x there is a bug with uninstalling packages on OpenSuse.
This is caused by the zypper package provider inheriting the rpm package provider
and not sets the uninstall_options correctly.
This causes the error: 'can't convert nil into Array'
By forcing it to use the rpm provider for removal we avoid this whole situation.