testing: overhaul travis setup, platforms, etc.
Travis has build stages now, which we should use to avoid clogging the
build pipeline with heavy acceptance suites that are predicated upon the
intake tests running.

Debian 7 was broken, this fixes it.

This also starts moving away from encapsulating everything within the
Makefile, since we have more programmatic control in the Rakefile.